* Exposing wrapper for getaddrinfo. * Adding more fields and new type for gai flags. * Renamed to GetAddressInfo due to Windows macros. * Added scope mutators to FInternetAddr * Properly inject the scope ids on all results when using GetLocalAdapterAddresses * Fix the endianness for the raw ip mutators * Fix mapping on address conversion * New return types for GetAddressInformation
#Jira UENET-886, UE-62159, UE-62269, UE-62266, UE-62319
#rb: ryan.gerleve

[CL 4267341 by Jake Leonard in Dev-Networking branch]

+++ Engine/Source/Runtime/Sockets/Public/SocketTypes.h
@@ -5,7 +5,7 @@
 #include "CoreMinimal.h"
 
 /** Indicates the socket protocol of socket being used, typically for BSD Sockets */
-enum class ESocketProtocolFamily : uint32
+enum class ESocketProtocolFamily : uint8
 {
 	/** No protocol family specification. Typically defined as AF_UNSPEC */
 	None,
